requirements.txt
Flask==2.3.3
twilio==8.10.0
slack-sdk==3.26.1
python-dotenv==1.0.0
